Update and rename osinfo.cpp to sysinfo.cpp
